[Context] Change "size_t" to "uint32_t" for "counts"

+++ include/core/xe_cmdlist.hpp
@@ -726,7 +726,7 @@ namespace xe
         AppendLaunchMultipleFunctionsIndirect(
             uint32_t numFunctions,                          ///< [in] maximum number of functions to launch
             Function* phFunctions,                          ///< [in][range(0, numFunctions)] handles of the function objects
-            const size_t* pNumLaunchArguments,              ///< [in] pointer to device memory location that will contain the actual
+            const uint32_t* pNumLaunchArguments,            ///< [in] pointer to device memory location that will contain the actual
                                                             ///< number of launch arguments; value must be less-than or equal-to
                                                             ///< numFunctions
             const thread_group_dimensions_t* pLaunchArgumentsBuffer,///< [in][range(0, numFunctions)] pointer to device buffer that will
